Feature: main.js 컴포넌트 분리
