+++ package.json
@@ -1,7 +1,6 @@
 {
-  "homepage": "https://luwbe1.github.io/front_5th_chapter1-1/",
   "name": "front-5th-chapter1-1",
-  "private": true,
+  "private": false,
   "version": "0.0.0",
   "type": "module",
   "scripts": {
@@ -22,7 +21,7 @@
     "prepare": "husky",
     "gh-pages": "bash gh-pages-publish.sh",
     "predeploy": "npm run build",
-    "deploy": "gh-pages -d build"
+    "deploy": "gh-pages -d dist"
   },
   "lint-staged": {
     "*.{js,jsx,ts,tsx}": [
@@ -47,5 +46,6 @@
     "prettier": "^3.4.2",
     "vite": "^6.0.3",
     "vitest": "^2.1.8"
-  }
+  },
+  "homepage": "https://luwbe1.github.io/front_5th_chapter1-1/"
 }
