+++ src/components/footer.js
@@ -1,5 +1,7 @@
-export const Footer = () => /*html*/ `
+const Footer = () => /*html*/ `
         <footer class="bg-gray-200 p-4 text-center">
           <p>&copy; 2024 항해플러스. All rights reserved.</p>
         </footer>
 `;
+
+export default Footer;
